Update Logs
Fixed the clear board function when resetting the game. There is an error with the row deletion functionality though. It still works, but it would not do it for every row. Sometimes it would clear, sometimes it would not.  I have an idea on why it does that, and I will debug it later today, or in the next couple days.